introduced constructor from blockchaing config for BlockDateGenerator
